Cleanup and fix issues
